+++ packages/jest-core/src/testSchedulerHelper.ts
@@ -27,24 +27,25 @@ export function shouldRunInBand(
   }
 
   /*
-   * Run in band if we only have one test or one worker available, unless we
-   * are using the watch mode, in which case the TTY has to be responsive and
-   * we cannot schedule anything in the main thread. Same logic applies to
-   * watchAll.
+   * If we are using watch/watchAll mode, don't schedule anything in the main
+   * thread to keep the TTY responsive and to prevent watch mode crashes caused
+   * by leaks (improper test teardown).
+   */
+  if (watch || watchAll) {
+    return false;
+  }
+
+  /*
+   * Otherwise, run in band if we only have one test or one worker available.
    * Also, if we are confident from previous runs that the tests will finish
    * quickly we also run in band to reduce the overhead of spawning workers.
    * Finally, the user can provide the runInBand argument in the CLI to
-   * force running in band.
-   * https://github.com/facebook/jest/blob/700e0dadb85f5dc8ff5dac6c7e98956690049734/packages/jest-config/src/getMaxWorkers.js#L14-L17
+   * force running in band, which sets maxWorkers to 1 here:
+   * https://github.com/facebook/jest/blob/d106643a8ee0ffa9c2f11c6bb2d12094e99135aa/packages/jest-config/src/getMaxWorkers.ts#L27-L28
    */
-  const isWatchMode = watch || watchAll;
   const areFastTests = timings.every(timing => timing < SLOW_TEST_TIME);
   const oneWorkerOrLess = maxWorkers <= 1;
   const oneTestOrLess = tests.length <= 1;
-
-  if (isWatchMode) {
-    return oneWorkerOrLess || (oneTestOrLess && areFastTests);
-  }
 
   return (
     // When specifying a memory limit, workers should be used